+++ movement_faces.h
@@ -52,4 +52,5 @@
 #include "nanosec_face.h"
 #include "mars_time_face.h"
 #include "peek_memory_face.h"
+#include "periodic_face.h"
 // New includes go above this line.
