+++ watch-faces.mk
@@ -27,4 +27,5 @@ SRCS += \
   ./watch-faces/settings/nanosec_face.c \
   ./watch-faces/io/chirpy_demo_face.c \
   ./watch-faces/io/irda_upload_face.c \
+  ./watch-faces/complication/periodic_face.c \
 # New watch faces go above this line.
